+++ CHANGES.rst
@@ -123,6 +123,11 @@ API changes
 
 - ``astropy.convolution``
 
+  - The parameter names of ``Gaussian2DKernel`` are now same as that of
+    ``Gaussian2DModel``. The ``__init__`` method of Gaussian2DKernel gives a
+    deprecation warning if ``stddev`` is used instead of ``x_stddev`` and
+    ``y_stddev``. [#3605]
+
 - ``astropy.coordinates``
 
   - Removed compatibility layer for pre-v0.4 API. [#4447]